Data file for products CRUD tests
